# Conflicts:
#   admin/src/core/seeds.py
#   admin/src/web/templates/base_index.html

+++ admin/src/core/repositories/horse.py
@@ -1,5 +1,5 @@
 from flask import current_app
-from src.core.models.employee import Employee
+from src.core.repositories import employee as employee_repo
 from src.core.database import db
 from src.core.models.horse import Horse
 
@@ -13,8 +13,6 @@ def list_horses(search='', assigned_activities_ja=None, sort_by='name', directio
         )
     if assigned_activities_ja:
        query = query.filter(Horse.assigned_activities_ja == assigned_activities_ja)
-    else:
-        query = query
 
     items_per_page = current_app.config.get('ITEMS_PER_PAGE')
 
@@ -26,6 +24,15 @@ def list_horses(search='', assigned_activities_ja=None, sort_by='name', directio
     pagination_horses = query.paginate(page=page, per_page=items_per_page, error_out=False)
 
     return pagination_horses
+
+def get_activities():
+    return Horse.assigned_activities_ja.property.columns[0].type.enums
+
+def get_genders():
+    return Horse.gender.property.columns[0].type.enums
+
+def get_purchase_donation():
+    return Horse.purchase_donation.property.columns[0].type.enums
 
 def find_horse_by_name(name):
     horse = Horse.query.filter(Horse.name == name).first()
@@ -40,21 +47,43 @@ def get_horse(id):
     return horse
 
 def create_horse(**kwargs):
+    trainer_ids = kwargs.pop('trainer_ids', [])
+
     horse = Horse(**kwargs)
     db.session.add(horse)
+
+    for trainer_id in trainer_ids:
+        trainer = employee_repo.get_employee(trainer_id)
+        if trainer:
+            horse.association.append(trainer)
+
     db.session.commit()
     return horse
 
 def update_horse(horse_id, **kwargs):
     horse = Horse.query.filter(Horse.id == horse_id).first()
+    if not horse:
+        return False
+
+    trainer_ids = kwargs.pop('trainer_ids', [])
+
     for key, value in kwargs.items():
         setattr(horse, key, value)
+
+    horse.association = []
+    for trainer_id in trainer_ids:
+        trainer = employee_repo.get_employee(trainer_id)
+        if trainer:
+            horse.association.append(trainer)
+
     db.session.commit()
-    return horse
+    return True
 
 def delete_horse(horse_id):
     horse = Horse.query.filter(Horse.id == horse_id).first()
-    db.session.delete(horse)
-    db.session.commit()
-    return horse
+    if horse:
+        db.session.delete(horse)
+        db.session.commit()
+        return True
+    return False
